migrate helpers.js to ts
