+++ src/signing.js
@@ -19,7 +19,7 @@ import * as Crypto from 'node:crypto'
 import _ from 'lodash'
 
 import * as errors from './errors.ts'
-import { getScope, isArray, isNumber, isObject, isString, makeDateLong, makeDateShort, uriEscape } from './helpers.js'
+import { getScope, isNumber, isObject, isString, makeDateLong, makeDateShort, uriEscape } from './helpers.ts'
 
 const signV4Algorithm = 'AWS4-HMAC-SHA256'
 
@@ -43,7 +43,7 @@ function getCanonicalRequest(method, path, headers, signedHeaders, hashedPayload
   if (!isObject(headers)) {
     throw new TypeError('headers should be of type "object"')
   }
-  if (!isArray(signedHeaders)) {
+  if (!Array.isArray(signedHeaders)) {
     throw new TypeError('signedHeaders should be of type "array"')
   }
   if (!isString(hashedPayload)) {
