+++ src/type.ts
@@ -1,0 +1,3 @@
+export type Binary = string | Buffer
+
+export type Mode = 'COMPLIANCE' | 'GOVERNANCE'
